--- app/services/stats/report_generator/rejections_refusals.rb
@@ -1,9 +1,0 @@
-module Stats
-  module ReportGenerator
-    class RejectionsRefusals < Base
-      def reporter
-        Reports::RejectionsRefusals
-      end
-    end
-  end
-end
