--- app/services/stats/report_generator/submitted_claims.rb
@@ -1,9 +1,0 @@
-module Stats
-  module ReportGenerator
-    class SubmittedClaims < Base
-      def reporter
-        Reports::SubmittedClaims
-      end
-    end
-  end
-end
