+++ app/services/stats/stats_report_generator.rb
@@ -33,13 +33,9 @@ module Stats
     end
 
     def generate_new_report
-      generator_klass = {
-        management_information: ManagementInformationGenerator,
-        provisional_assessment: ReportGenerator::ProvisionalAssessment,
-        rejections_refusals: ReportGenerator::RejectionsRefusals,
-        submitted_claims: ReportGenerator::SubmittedClaims
-      }[report_type.to_sym]
-      generator_klass.call(options)
+      return ManagementInformationGenerator.call(options) if report_type.to_sym == :management_information
+
+      ReportGenerator.call(report_type, **options)
     end
 
     def handle_error(error, record)
